Height and Weight now are Float type, and putting something that a number doesn't make crash the app anymore. The app was trying to toFloat une String non filtré.

+++ app/src/main/java/com/example/k_fit/presentation/features/register/RegisterProfileState.kt
@@ -16,6 +16,6 @@ data class RegisterProfileState(
     var lastName: String = "",
     var birthDate: String = "",
     var gender: Gender? = null,
-    var weight: Int = 0,
-    var height: Int = 0,
+    var weight: Float = 0f,
+    var height: Float = 0f,
 )
